Initial update to implementation for new endpoint interface.
The code compiles "clean" for BGP, BGQ, and MPI, with many parts of
the code if 0'd out to get a compile.

The socklinux target needs more work (DTimer and Datagram problems)

+++ common/bgp/Client.h
@@ -89,10 +89,10 @@ namespace XMI
           return _name;
         }
 
-	inline xmi_result_t createContext_impl (xmi_configuration_t configuration[],
-					size_t count,
-					xmi_context_t *context,
-					int *ncontexts) {
+	inline xmi_result_t createContext_impl (xmi_configuration_t   configuration[],
+					        size_t                count,
+					        xmi_context_t       * context,
+					        size_t              * ncontexts) {
 		//_context_list->lock ();
 		int n = *ncontexts;
 		if (_ncontexts != 0) {
